requirements.txt
speedtest-cli
pyexcel
pyexcel-xlsx
bitmath